fix: /home to /
