+++ client/src/app/(user)/play/friend/page.tsx
@@ -210,7 +210,7 @@ export default function PlayFriend() {
 
         socketInstance.on("opponentBanned", () => {
           toast.info("Admin banned your opponent.");
-          router.push("/home");
+          router.push("/");
         });
 
         socketInstance.on("playRequestReceived", (data) => {
