+++ client/src/app/not-found.tsx
@@ -10,7 +10,7 @@ function NotFoundPage() {
       <div className="bg-[#FF6A3D] px-2 text-sm rounded rotate-12 absolute">
         Page Not Found
       </div>
-      <Link href={"/home"}>
+      <Link href={"/"}>
         <button className="mt-5">
           <a className="relative inline-block text-sm font-medium text-[#FF6A3D] group active:text-orange-500 focus:outline-none focus:ring">
             <span className="absolute inset-0 transition-transform translate-x-0.5 translate-y-0.5 bg-[#FF6A3D] group-hover:translate-y-0 group-hover:translate-x-0"></span>
