+++ client/src/components/ClientSocketInitializer.tsx
@@ -1,14 +1,38 @@
 "use client";
 
 import { useEffect } from "react";
-import { useFriendStore } from "@/stores/useFriendStore"; 
+import { usePathname, useRouter } from "next/navigation";
+import { toast } from "sonner";
+import { useAuthStore } from "@/stores";
+import { useFriendStore } from "@/stores/useFriendStore";
 
 export default function ClientSocketInitializer() {
-  const initializeSocket = useFriendStore((state) => state.initializeSocket);
+  const { user, logout } = useAuthStore();
+  const { initializeSocket } = useFriendStore();
+  const router = useRouter();
+  // const pathname = usePathname();
 
   useEffect(() => {
-    initializeSocket();
-  }, [initializeSocket]);
+    if (user?._id) {
+      const socket = initializeSocket();
+      if (socket) {
+        socket.emit("join", user._id);
+
+        socket.on("userBanned", (data) => {
+          if (data.userId === user._id) {
+            toast.error("You have been banned by an admin.");
+            logout();
+            router.push("/login");
+          }
+        });
+
+        return () => {
+          socket.off("userBanned");
+        };
+      }
+    }
+  }, [user?._id, initializeSocket, router, logout]);
+
 
   return null;
 }
